+++ packages/seed/src/index.ts
@@ -1,0 +1,3 @@
+export { LocalConfig } from './localConfig'
+export { FixtureStorage } from './fixtureStorage'
+export { config } from './config'