+++ packages/services-http/src/content-type/index.ts
@@ -1,0 +1,2 @@
+export { HttpError } from '../typings/http_error'
+export { HttpJson } from './http_json'