+++ packages/services-http/src/typings/http_error.ts
@@ -1,0 +1,7 @@
+import { Typings } from '@seagull/libraries'
+import { Response } from 'node-fetch'
+
+export interface HttpError extends Typings.BasicError {
+  message: string
+  details: Response
+}