+++ packages/services-s3/src/mode/edge.ts
@@ -1,0 +1,12 @@
+import { IMode } from '@seagull/mode'
+import { injectable } from 'inversify'
+import 'reflect-metadata'
+import { S3Base } from './base'
+
+/**
+ * Http edge mode implementation.
+ */
+@injectable()
+export class S3Edge extends S3Base {
+  protected mode: Readonly<IMode> = { environment: 'edge' }
+}